물려주기 post

+++ booksale/src/main/java/dbteam4/booksale/dto/PostDTO.java
@@ -17,7 +17,7 @@ public class PostDTO {
     private String postTitle;
     private String postContent;
     private String salesMethod;
-    private int postPrice;
+    private Long postPrice;
     private String shippingMethod;
     private String bookQuality;
     private LocalDateTime postTime;
